requirements.txt
azely
jinja2
requests
python-frontmatter
python-slugify
shapely
numpy~=1.26.4
timezonefinder
python-dateutil
pyyaml
